# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Weather Station API",
    "description": "JSON REST API for recording weather station data"
  },
  "servers": [
    {
      "url": "http://localhost:8081/"
    }
  ],
  "paths": {
...
    "/readings": {
      "post": {
        "summary": "Insert a new reading",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Insert a new weather reading document",
          "content": {
            "application/json": {
              "schema": {
                "deviceName": "string",
                "precipitation": "number",
                "time": "string",
                "latitude": "number",
                "longitude": "number",
                "atmosphericPressure": "number",
                "maxWindSpeed": "number",
                "solarRadiation": "number",
                "vaporPressure": "number",
                "humidity": "number",
                "temperature": "number",
                "windDirection": "number"
              },
              "example": {
                "deviceName": "My_Sensor",
                "precipitation": "0.085",
                "time": "2021-05-07T02:44:05.000+00:00",
                "latitude": "152.77891",
                "longitude": "-26.95064",
                "atmosphericPressure": "128.07",
                "maxWindSpeed": "4.16",
                "solarRadiation": "493.42",
                "vaporPressure": "1.8",
                "humidity": "72.76",
                "temperature": "23.89",
                "windDirection": "153.6"
              }
            }
          }
        }
      },
      "put": {
        "summary": " Update one readings specific entry",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update a single readings precipitation value",
          "content": {
            "application/json": {
              "schema": {
                "id": "string",
                "precipitation": "number"
              },
              "example": {
                "id": "642e60ef9c4141e1fd3c897a",
                "precipitation": "0.9"
              }
            }
          }
        }
      }
    },
    "/readings-many": {
      "post": {
        "summary": "Insert Many Readings as an array",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "description": "Insert multiple (plural) readings for a single station",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "deviceName": "string",
                  "readings": "array",
                  "items": {
                    "precipitation": "number",
                    "time": "string",
                    "latitude": "number",
                    "longitude": "number",
                    "atmosphericPressure": "number",
                    "maxWindSpeed": "number",
                    "solarRadiation": "number",
                    "vaporPressure": "number",
                    "humidity": "number",
                    "temperature": "number",
                    "windDirection": "number"
                  }
                }
              },
              "example": {
                "deviceName": "My_Sensor",
                "readings": [
                  {
                    "precipitation": "0.085",
                    "time": "2021-05-07T02:44:05.000+00:00",
                    "latitude": "152.77891",
                    "longitude": "-26.95064",
                    "atmosphericPressure": "128.07",
                    "maxWindSpeed": "4.16",
                    "solarRadiation": "493.42",
                    "vaporPressure": "1.8",
                    "humidity": "72.76",
                    "temperature": "23.89",
                    "windDirection": "153.6"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/readings/max-precipitation": {
      "get": {
        "summary": "Get list of max precipitation from readings",
        "description": "",
        "parameters": [
          {
            "name": "deviceName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {}
      }
    },
...
    "/readings/max-temp": {
      "get": {
        "summary": "Get all readings Max Temperature by date/time",
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Get Max Temp by date/time range for all sensors,  returning only sensor name, reading date/time and the precipitation values",
          "content": {
            "application/json": {
              "schema": {
                "startDate": "string",
                "endDate": "string"
              },
              "example": {
                "startDate": "2020-01-01",
                "endDate": "2023-01-01"
              }
            }
          }
        }
      }
    }
  }
}